refactor(core-theme): update structure
