+++ packages/core/theme/index.ts
@@ -1,0 +1,28 @@
+export {
+    ThemeSchemeState,
+    ThemeSchemeType,
+} from './src/interfaces/theme-scheme.state';
+
+export { ColorSchemeService, ColorSchemeType } from './src/services/color-scheme.service';
+export { StyleManagerService } from './src/services/style-manager.service';
+export { ThemeColorService } from './src/services/theme-color.service';
+export {
+    ThemeSchemeService,
+} from './src/services/theme-scheme.service';
+export {
+    DarkThemeManagerConfig,
+    XmDarkThemeManager,
+} from './src/services/xm-dark-theme-manager';
+export {
+    themeInitializerFactory,
+    THEME_PROVIDER_FACTORY,
+    themeInitializer,
+    XmThemeConfig,
+    XmThemeLoader,
+} from './src/services/xm-theme-loader.service';
+
+export {
+    THEME_STRATEGY,
+    XmTheme,
+    XmThemeStore,
+} from './src/stores/xm-theme-store.service';
