--- packages/theme/public_api.ts
@@ -1,26 +1,0 @@
-export { ColorSchemeService, ColorSchemeType } from './src/color-scheme.service';
-export { StyleManagerService } from './src/style-manager.service';
-export { ThemeColorService } from './src/theme-color.service';
-export {
-    ThemeSchemeService,
-} from './src/theme-scheme.service';
-export {
-    ThemeSchemeState,
-    ThemeSchemeType,
-} from './src/theme-scheme.state';
-export {
-    DarkThemeManagerConfig,
-    XmDarkThemeManager,
-} from './src/xm-dark-theme-manager';
-export {
-    themeInitializerFactory,
-    THEME_PROVIDER_FACTORY,
-    themeInitializer,
-    XmThemeConfig,
-    XmThemeLoader,
-} from './src/xm-theme-loader.service';
-export {
-    THEME_STRATEGY,
-    XmTheme,
-    XmThemeStore,
-} from './src/xm-theme-store.service';
