+++ tsconfig.json
@@ -32,7 +32,7 @@
       "@xm-ngx/core/global-error-handler": ["packages/core/global-error-handler"],
       "@xm-ngx/core/permission": ["src/app/shared/privilege"],
       "@xm-ngx/core/permission/testing": ["packages/core/permission/testing"],
-      "@xm-ngx/core/theme": ["packages/theme/public_api"],
+      "@xm-ngx/core/theme": ["packages/core/theme"],
       "@xm-ngx/core/environment": ["src/environments"],
       "@xm-ngx/ext/*": ["src/app/ext/*"],
       "@xm-ngx/alert": ["packages/alert"],
